refactor(router): replace LoadInstructionResolver with routing instruction and router
